Moved capture.js to ts and WIP adding useElectronCaptureAPI() and useAlternativeCaptureAPI()
